% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvML.R
\docType{package}
\name{deconvML}
\alias{deconvML}
\title{deconvML: A single cell data resource}
\description{
The deconvML package allows to build a catalog that helps the user to answer
some questions about their dataset through the following functions:
getDataSetList, getDataSetInfo, getCellTypesInDataSet, getActiveGeneMethod,
getTissueTypesInDataSet and getActiveGenesInCellType.
}
\section{deconvML functions}{

getDataSetList List data sets included.

getDataSetInfo List main properties describing a data set.

getCellTypesInDataSet List cell types included in a specific data set.

getActiveGeneMethod List available methods to estimate the active genes in a
specific data set.

getTissueTypesInDataSet List tissue types included in a specific data set

getActiveGenesInCellType List active genes in the tissue, cell type, of a
specific data set for a given method
}

