% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getActiveGeneMethod}
\alias{getActiveGeneMethod}
\title{List available methods to estimate the active genes in a specific data set}
\usage{
getActiveGeneMethod(dataset)
}
\arguments{
\item{dataset}{String with the name of the data set.}
}
\value{
available methods data.
}
\description{
getActiveGeneMethod returns the methods available to apply on a
  specific data set.
}
