% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getActiveGenesInCellType}
\alias{getActiveGenesInCellType}
\title{List active genes in the tissue, cell type, of a specific data set for a
given method}
\usage{
getActiveGenesInCellType(dataset, tissue, celltype, method)
}
\arguments{
\item{dataset}{String with the name of the data set.}

\item{celltype}{String with the name of the celltype in which we are
interested.}

\item{method}{String with the name of the method in which we are interested.}

\item{tisue}{String with the name of the tissue in which we are interested.}
}
\value{
gene ids of genes active in the tissue, cell type, for a given
  calculus method.
}
\description{
getActiveGenesInCellTypes read the files containing the
  pre-calculated matrices to return ids of active genes in the tissue, cell
  type of a specific data set for a given method.
}
