% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getCellTypesInDataSet}
\alias{getCellTypesInDataSet}
\title{List cell types included in a specific data set}
\usage{
getCellTypesInDataSet(dataset)
}
\arguments{
\item{dataset}{String with the name of the data set.}
}
\value{
cell types considered in the data set identified by the string.
}
\description{
getCellTypesInDataSet shows the different cell types for a
  specific data set.
}
