% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getDataSetInfo}
\alias{getDataSetInfo}
\title{List main properties describing a data set}
\usage{
getDataSetInfo(dataset)
}
\arguments{
\item{dataset}{String with the name of the data set.}
}
\value{
data frame of properties describing the dataset identified by the
  string.
}
\description{
getDataSetInfo displays the following characteristics of a data
  set: version, number of cells, number of genes, number of type of tissues
  and number of type of cells.
}
