% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getDataSetList}
\alias{getDataSetList}
\title{List data sets included}
\usage{
getDataSetList()
}
\value{
vector of strings with the data sets included.
}
\description{
getDataSetList shows data sets included as they are loom class variables.
}
