% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getTissueTypesInDataSet}
\alias{getTissueTypesInDataSet}
\title{List tissue types included in a specific data set}
\usage{
getTissueTypesInDataSet(dataset)
}
\arguments{
\item{dataset}{String with the name of the data set.}
}
\value{
the tissue types considered in the data set identified by the string.
}
\description{
getTissueTypesInDataSet collects all types of tissues included
  in the data set.
}
